Remove explicit `apiVersion`, as this is now corrected in ksonnet-lib
This commit will remove a work-around for `ksonnet-lib`'s issue 44. It
has no effect on the generated JSON.

`ksonnet-lib` issue 44 describes an issue that the group is sometimes
not set correctly when we generate the library. In particular,
Kubernetes is moving towards a model where a group is identified in the
OpenAPI spec by a domain like `rbac.authorization.k8s.io`, rather than a
simple identifier (e.g., `rbac`) as was true previously.

Since this issue will be fixed in `ksonnet-lib`'s PR 46, we need not
explicitly correct the `apiVersion` field.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>